requirements.txt
absl-py==2.1.0
accelerate==1.0.0
aiofiles==23.2.1
aiohttp==3.9.3
aiosignal==1.3.1
altair==5.2.0
altgraph==0.17.4
annotated-types==0.6.0
anthropic==0.2.10
anyio==4.2.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
args==0.1.0
arrow==1.3.0
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.2.0
audioread==3.0.1
azure-cognitiveservices-speech==1.36.0
Babel==2.14.0
beartype==0.17.2
beautifulsoup4==4.12.3
bibtexparser==2.0.0b7
bleach==6.1.0
blinker==1.9.0
Brotli==1.1.0
cachetools==5.3.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
clldutils==3.22.1
colorama==0.4.6
colorlog==6.8.2
comm==0.2.1
contourpy==1.2.0
cryptography==44.0.0
csvw==3.3.0
ctransformers==0.2.27
cycler==0.12.1
dataclasses-json==0.5.14
debugpy==1.8.1
decorator==4.4.2
deepl_cli==0.6.0
defusedxml==0.7.1
Deprecated==1.2.14
deprecation==2.1.0
diffusers==0.30.3
diskcache==5.6.3
display==1.0.0
distro==1.9.0
dlinfo==1.2.1
docutils==0.21.2
e==1.4.5
einops==0.7.0
elevenlabs==0.3.0b0
exceptiongroup==1.2.0
executing==2.0.1
fastapi==0.100.1
fastjsonschema==2.19.1
ffmpeg==1.4
ffmpeg-python==0.2.0
ffmpy==0.3.2
filelock==3.13.1
Flask==3.1.0
Flask-Cors==5.0.0
flatbuffers==24.3.25
fonttools==4.49.0
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.2.0
future==1.0.0
gast==0.6.0
generativeai==0.0.1
gitdb==4.0.11
GitPython==3.1.43
gmpy2==2.1.2
google-ai-generativelanguage==0.4.0
google-api-core==2.17.1
google-auth==2.28.0
google-generativeai==0.3.2
google-pasta==0.2.0
googleapis-common-protos==1.62.0
gradio==4.15.0
gradio_client==0.8.1
greenlet==3.0.3
grpcio==1.60.1
grpcio-status==1.60.1
h11==0.14.0
h5py==3.12.1
hjson==3.1.0
httpcore==1.0.3
httpx==0.27.2
huggingface-hub==0.25.2
idna==3.10
imageio==2.34.1
imageio-ffmpeg==0.5.1
img2pdf==0.5.1
importlib-resources==6.1.1
importlib_metadata==8.5.0
inflect==7.0.0
install-playwright==0.0.0
intel-cmplr-lib-ur==2025.0.0
intel-openmp==2025.0.0
ipykernel==6.29.2
ipython==8.21.0
ipywidgets==8.1.2
isodate==0.6.1
isoduration==20.11.0
itsdangerous==2.2.0
janus==1.0.0
jax==0.4.35
jaxlib==0.4.35
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
json5==0.9.17
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter-console==6.6.3
jupyter-events==0.9.0
jupyter-lsp==2.2.2
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyter_server==2.12.5
jupyter_server_terminals==0.5.2
jupyterlab==4.1.2
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.3
jupyterlab_widgets==3.0.10
keras==3.6.0
kiwisolver==1.4.5
langchain==0.0.198
langchainplus-sdk==0.0.20
language-tags==1.2.0
lazy_loader==0.3
lerc==4.0.1
libclang==18.1.1
libpcap==1.11.0b12
librosa==0.10.1
llama_cpp_python==0.3.0
llvmlite==0.42.0
lxml==5.1.0
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.21.1
matplotlib==3.8.3
matplotlib-inline==0.1.6
mdurl==0.1.2
mediapipe==0.10.14
miniaudio==1.59
mistune==3.0.2
mkl-fft==1.3.8
mkl-random==1.2.4
mkl-service==2.4.0
ml-dtypes==0.4.1
more-itertools==10.5.0
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.0.8
mss==9.0.1
multidict==6.0.5
mypy-extensions==1.0.0
namex==0.0.8
nbclient==0.9.0
nbconvert==7.16.1
nbformat==5.9.2
nest-asyncio==1.6.0
networkx==3.1
nltk==3.8.1
notebook==7.1.0
notebook_shim==0.2.4
numba==0.59.0
numexpr==2.9.0
numpy==1.26.3
ocrmypdf==16.6.2
ollama==0.3.3
openai==0.27.10
openapi-schema-pydantic==1.2.4
opencv-contrib-python==4.10.0.84
opencv-python==4.9.0.80
opentelemetry-api==1.23.0
opentelemetry-sdk==1.23.0
opentelemetry-semantic-conventions==0.44b0
opt_einsum==3.4.0
optree==0.13.0
orjson==3.9.14
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.1
pandas==2.2.0
pandocfilters==1.5.1
parso==0.8.3
pdfminer.six==20240706
pefile==2023.2.7
phonemizer==3.2.1
pi_heif==0.20.0
pikepdf==9.4.2
pillow==10.4.0
pkg-about==1.2.1
platformdirs==4.2.0
playwright==1.41.2
pluggy==1.5.0
pooch==1.8.1
proglog==0.1.10
progressbar==2.5
prometheus_client==0.20.0
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.3
psutil==5.9.8
pure-eval==0.2.2
py-cpuinfo==9.0.0
py-espeak-ng==0.1.8
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyAudio==0.2.14
pycparser==2.21
pycryptodome==3.21.0
pydantic==1.9.2
pydantic_core==2.23.4
pydub==0.25.1
pyee==11.0.1
pygame==2.6.0
Pygments==2.17.2
pyinstaller==6.11.0
pyinstaller-hooks-contrib==2024.9
pylatexenc==2.10
PyOpenGL==3.1.7
PyOpenGL-accelerate==3.1.7
pyparsing==3.1.1
PyPDF2==3.0.1
PySocks==1.7.1
pytesseract==0.3.13
python-dateutil==2.8.2
python-json-logger==2.0.7
python-multipart==0.0.6
pytz==2024.1
pywin32==306
pywin32-ctypes==0.2.3
pywinpty==2.0.12
PyYAML==6.0.1
pyzmq==25.1.2
qtconsole==5.5.1
QtPy==2.4.1
rdflib==7.0.0
referencing==0.33.0
regex==2023.12.25
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986==1.5.0
rfc3986-validator==0.1.1
rich==13.7.0
rotary-embedding-torch==0.5.3
rpds-py==0.18.0
rsa==4.9
ruff==0.2.2
safetensors==0.4.5
scapy==2.5.0
scikit-learn==1.4.1.post1
scipy==1.12.0
seaborn==0.13.2
segments==2.2.1
selenium==4.26.1
semantic-version==2.10.0
Send2Trash==1.8.2
sentencepiece==0.2.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
sortedcontainers==2.4.0
sounddevice==0.4.6
soundfile==0.12.1
soupsieve==2.5
soxr==0.3.7
SpeechRecognition==3.10.1
SQLAlchemy==2.0.28
stack-data==0.6.3
starlette==0.27.0
sympy==1.12
tabulate==0.9.0
tasklib==2.5.1
tcmlib==1.2.0
tenacity==8.2.3
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-intel==2.17.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.5.0
terminado==0.18.0
tesseract==0.1.3
threadpoolctl==3.3.0
tiktoken==0.3.1
timm==0.9.12
tinycss2==1.2.1
tk==0.1.0
tokenizers==0.13.3
tomli==2.0.2
tomlkit==0.12.0
toolz==0.12.1
torch==2.4.0+cu121
torchaudio==2.4.0+cu121
torchvision==0.19.0+cu121
tornado==6.4
tqdm==4.66.6
traitlets==5.14.1
transformers==4.31.0
trio==0.27.0
trio-websocket==0.11.1
typer==0.9.0
types-python-dateutil==2.8.19.20240106
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2024.1
ultralytics==8.3.24
ultralytics-thop==2.0.10
umf==0.9.0
Unidecode==1.3.8
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.2.3
urwid==2.6.8
uvicorn==0.21.1
vit==2.3.2
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.8.0
websockets==11.0.3
Werkzeug==3.1.3
whisper==1.1.10
widgetsnbextension==4.0.10
win-inet-pton==1.1.0
WMI==1.5.1
wrapt==1.16.0
wsproto==1.2.0
xformers==0.0.27.post2
yarl==1.9.4
youtube-dl==2021.12.17
zipp==3.20.2
zstd==1.5.5.1
